+++ test/ElasticsearchQueryLucene.FunctionalTests/UnitTest1.cs
@@ -1,0 +1,10 @@
+﻿namespace ElasticsearchQueryLucene.FunctionalTests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+
+    }
+}